moved the entities.hpp file to the include directory
