+++ src/engine/physics.cpp
@@ -1,4 +1,4 @@
-#include "entities.hpp"
+#include "physics.hpp"
 
 void gravity(Player *player) { player->dstrect.y += player->accel; }
